Refactoring:
Moved Note-Traits and implementations to distinct files seperated by modules to declutter notes module
