+++ lib/src/notes/mod.rs
@@ -1,0 +1,6 @@
+pub mod traits;
+pub mod grouped_remote_note_headers;
+pub mod remote_note_metadata;
+pub mod remote_note_header_collection;
+pub mod note_headers;
+pub mod localnote;
