+++ lib/src/notes/traits/mod.rs
@@ -1,0 +1,3 @@
+pub mod identifyable_note;
+pub mod mergeable_note_body;
+pub mod header_parser;
